added aditional prop to rich text editor
