+++ src/components/RichTextEditor/RichTextEditor.stories.tsx
@@ -4,8 +4,9 @@ import { Meta, Story } from '@storybook/react';
 import React from 'react';
 import { RichTextEditor as RichTextEditorComponent, RichTextEditorProps } from './RichTextEditor';
 import { serializeNodesToHtml } from './serializer/serializeToHtml';
+import { Position } from './types';
 import { EditorActions } from './utils/actions';
-import { checkboxValue, customDesignTokens, htmlValue, IPSUM, nodesToSerialize, value } from './utils/exampleValues';
+import { IPSUM, checkboxValue, customDesignTokens, htmlValue, nodesToSerialize, value } from './utils/exampleValues';
 
 export default {
     title: 'Components/Rich Text Editor',
@@ -15,11 +16,13 @@ export default {
         placeholder: 'Some placeholder',
         readonly: false,
         clear: false,
+        position: Position.FLOATING,
     },
     argTypes: {
         onTextChange: { action: 'onTextChange' },
         onBlur: { action: 'onBlur' },
-        vale: { type: 'string' },
+        value: { type: 'string' },
+        position: { options: [Position.FLOATING, Position.BOTTOM] },
     },
 } as Meta;
