+++ src/components/RichTextEditor/components/Toolbar/types.d.ts
@@ -1,0 +1,8 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
+import { EditorActions } from '../../utils/actions';
+
+type ToolbarBottomProps = {
+    editorId?: string;
+    actions?: EditorActions[][];
+};
