+++ src/components/RichTextEditor/types.ts
@@ -26,3 +26,8 @@ export type ButtonGroupProps = {
 export type ButtonGroupWidths = { actions: EditorActions[]; buttonGroupWidth: number; index: number }[];
 
 export type AvailableTextStyles = ListStyles & TextStyles;
+
+export enum Position {
+    FLOATING = 'floating',
+    BOTTOM = 'bottom',
+}
