+++ src/components/RichTextEditor/utils/constants.ts
@@ -1,0 +1,3 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
+export const ON_SAVE_DELAY_IN_MS = 500;
